+++ java-checks-test-sources/pom.xml
@@ -700,12 +700,6 @@
       <version>4.0.10</version>
       <scope>provided</scope>
     </dependency>
-    <dependency>
-      <groupId>au.com.dius</groupId>
-      <artifactId>pact-jvm-provider-spring_2.12</artifactId>
-      <version>3.6.15</version>
-      <scope>provided</scope>
-    </dependency>
   </dependencies>
 
   <profiles>
