+++ java-checks-test-sources/src/main/java/checks/tests/NoTestInTestClassCustomPattern.java
@@ -1,4 +1,4 @@
-package checks;
+package checks.tests;
 
 public class NoTestInTestClassCustomPattern {
 }
